+++ src/main/java/terramine/client/render/entity/EaterOfSoulsRenderer.java
@@ -36,7 +36,7 @@ public class EaterOfSoulsRenderer extends MobRenderer<EaterOfSoulsEntity, EaterO
 
     @Override
     protected void setupRotations(@NotNull EaterOfSoulsEntity entity, @NotNull PoseStack poseStack, float f, float g, float h) {
-        poseStack.mulPose(Vector3f.XP.rotationDegrees(entity.getXRot()));
         super.setupRotations(entity, poseStack, f, g, h);
+        poseStack.mulPose(Vector3f.XP.rotationDegrees(entity.getXRot()));
     }
 }