+++ src/main/java/terramine/common/world/biome/CorruptionBiome.java
@@ -15,7 +15,7 @@ public class CorruptionBiome {
     private static Biome createCorruption() {
         MobSpawnSettings.Builder spawnSettings = new MobSpawnSettings.Builder();
         //BiomeDefaultFeatures.monsters(spawnSettings, 100, 25, 100, false);
-        spawnSettings.addSpawn(MobCategory.MONSTER, new MobSpawnSettings.SpawnerData(ModEntities.EATER_OF_SOULS, 50, 1, 2));
+        spawnSettings.addSpawn(MobCategory.MONSTER, new MobSpawnSettings.SpawnerData(ModEntities.EATER_OF_SOULS, 25, 0, 2));
 
         BiomeGenerationSettings.Builder generationSettings = new BiomeGenerationSettings.Builder();
         BiomeDefaultFeatures.addDefaultCarversAndLakes(generationSettings);